Organize project directory into typical python style
