--- settings.conf
@@ -1,2 +1,0 @@
-debug=true
-keep=false
