+++ will.py
@@ -1,0 +1,4 @@
+import will
+
+
+will.run()
