+++ will/ftools.py
