+++ will/logger/__init__.py
@@ -1,5 +1,5 @@
 import logging
-from logger.log_config import init_logger
+from will.logger.log_config import init_logger
 
 log = logging.getLogger('WILL')
