+++ will/logger/log_config.py
@@ -1,10 +1,10 @@
 import logging
 import os
 import sys
-import config
+import will.config as config
 from voluptuous import Schema, Required, Invalid, MultipleInvalid
-from logger.fallback import error
-from logger.handlers import ColorStreamHandler
+from will.logger.fallback import error
+from will.logger.handlers import ColorStreamHandler
 
 
 class LogConfig(object):
